refactor: recent 하위 페이지들의 새로운 폴더 경로를 구축합니다.
