+++ apps/client/components/Recent/TagFamily/TagFamily.constants.ts
@@ -1,15 +1,18 @@
 const tagList = [
   {
     id: "restaurant",
-    name: "Restaurant"
+    name: "Restaurant",
+    path: "/filter/restaurant"
   },
   {
     id: "nationality",
-    name: "Nationality"
+    name: "Nationality",
+    path: "/filter/nationality"
   },
   {
     id: "eatingHabit",
-    name: "Eating habit"
+    name: "Eating habit",
+    path: "/filter/eatingHabit"
   }
 ] as const;
