+++ apps/client/components/Recent/index.ts
@@ -1,6 +1,6 @@
+export { default as FilterEatingHabitPage } from "./FilterEatingHabitPage";
+export { default as FilterNationalityPage } from "./FilterNationalityPage";
+export { default as FilterRestaurantPage } from "./FilterRestaurantPage";
 export { default as History } from "./History";
-export { default as TagFamily } from "./TagFamily";
 export { default as RecentPage } from "./RecentPage";
-export { default as RecentRestaurantPage } from "./RecentRestaurantPage";
-export { default as RecentNationalityPage } from "./RecentNationalityPage";
-export { default as RecentEatingHabitPage } from "./RecentEatingHabitPage";
+export { default as TagFamily } from "./TagFamily";
