--- apps/client/pages/recent/eatingHabit.page.tsx
@@ -1,3 +1,0 @@
-import { RecentEatingHabitPage } from "~/components/Recent";
-
-export default RecentEatingHabitPage;
