+++ apps/client/pages/recent/filter/eatingHabit.page.tsx
@@ -1,0 +1,3 @@
+import { FilterEatingHabitPage } from "~/components/Recent";
+
+export default FilterEatingHabitPage;
