+++ apps/client/pages/recent/filter/nationality.page.tsx
@@ -1,0 +1,3 @@
+import { FilterNationalityPage } from "~/components/Recent";
+
+export default FilterNationalityPage;
