+++ apps/client/pages/recent/filter/restaurant.page.tsx
@@ -1,0 +1,3 @@
+import { FilterRestaurantPage } from "~/components/Recent";
+
+export default FilterRestaurantPage;
