--- apps/client/pages/recent/nationality.page.tsx
@@ -1,3 +1,0 @@
-import { RecentNationalityPage } from "~/components/Recent";
-
-export default RecentNationalityPage;
