--- apps/client/pages/recent/restaurant.page.tsx
@@ -1,3 +1,0 @@
-import { RecentRestaurantPage } from "~/components/Recent";
-
-export default RecentRestaurantPage;
